+++ packages/datasource-sql/src/connection/index.ts
@@ -1,67 +1,31 @@
-import type { ConnectionOptions, SslMode } from '../types';
+import type { ConnectionOptions, ConnectionOptionsObj } from '../types';
 import type { Logger } from '@forestadmin/datasource-toolkit';
 
-import { Dialect, Sequelize } from 'sequelize';
+import { Sequelize } from 'sequelize';
 
+import handleErrors from './handle-errors';
 import preprocessOptions from './preprocess';
 import ReverseProxy from './reverse-proxy';
-import { getLogger, getSchema, handleSequelizeError } from './utils';
-
-function getSslConfiguration(
-  dialect: Dialect,
-  sslMode: SslMode,
-  logger?: Logger,
-): Record<string, unknown> {
-  switch (dialect) {
-    case 'mariadb':
-      if (sslMode === 'disabled') return { ssl: false };
-      if (sslMode === 'required') return { ssl: { rejectUnauthorized: false } };
-      if (sslMode === 'verify') return { ssl: true };
-      break;
-
-    case 'mssql':
-      if (sslMode === 'disabled') return { options: { encrypt: false } };
-      if (sslMode === 'required')
-        return { options: { encrypt: true, trustServerCertificate: true } };
-      if (sslMode === 'verify')
-        return { options: { encrypt: true, trustServerCertificate: false } };
-      break;
-
-    case 'mysql':
-      if (sslMode === 'disabled') return { ssl: false };
-      if (sslMode === 'required') return { ssl: { rejectUnauthorized: false } };
-      if (sslMode === 'verify') return { ssl: { rejectUnauthorized: true } };
-      break;
-
-    case 'postgres':
-      if (sslMode === 'disabled') return { ssl: false };
-      if (sslMode === 'required') return { ssl: { require: true, rejectUnauthorized: false } };
-      if (sslMode === 'verify') return { ssl: { require: true, rejectUnauthorized: true } };
-      break;
-
-    case 'db2':
-    case 'oracle':
-    case 'snowflake':
-    case 'sqlite':
-    default:
-      if (sslMode && sslMode !== 'manual') {
-        logger?.('Warn', `ignoring sslMode=${sslMode} (not supported for ${dialect})`);
-      }
-
-      return {};
-  }
-}
+import { getLogger, getSchema, getSslConfiguration } from './utils';
 
 /** Attempt to connect to the database */
 export default async function connect(
   uriOrOptions: ConnectionOptions,
   logger?: Logger,
 ): Promise<Sequelize> {
-  let proxy: ReverseProxy | undefined;
-  let sequelize: Sequelize | undefined;
+  let proxy: ReverseProxy;
+  let sequelize: Sequelize;
+  let options: ConnectionOptionsObj;
+  let servername: string;
 
   try {
-    let options = await preprocessOptions(uriOrOptions);
+    options = await preprocessOptions(uriOrOptions);
+
+    try {
+      servername = options?.host ?? new URL(options.uri).hostname;
+    } catch {
+      // don't crash if using unix socket, sqlite, etc...
+    }
 
     if (options.proxySocks) {
       proxy = new ReverseProxy(options);
@@ -75,7 +39,7 @@ export default async function connect(
 
     opts.dialectOptions = {
       ...(opts.dialectOptions ?? {}),
-      ...getSslConfiguration(opts.dialect, sslMode, logger),
+      ...getSslConfiguration(opts.dialect, sslMode, servername, logger),
     };
 
     sequelize = uri
@@ -97,6 +61,6 @@ export default async function connect(
   } catch (e) {
     await sequelize?.close();
     // if proxy encountered an error, we want to throw it instead of the sequelize error
-    handleSequelizeError(proxy?.error || (e as Error));
+    handleErrors(proxy?.error || e, options, proxy);
   }
 }