+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,43 @@
+## example [1.3.13](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.12...example@1.3.13) (2023-05-31)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.6.4
+
+## example [1.3.12](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.11...example@1.3.12) (2023-05-31)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.2
+
+## example [1.3.11](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.10...example@1.3.11) (2023-05-31)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.6.3
+
+## example [1.3.10](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.9...example@1.3.10) (2023-05-30)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.1
+
 ## example [1.3.10-alpha-webhook-custom-actions.2](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.10-alpha-webhook-custom-actions.1...example@1.3.10-alpha-webhook-custom-actions.2) (2023-05-31)